feat(playground): refactoring file uploader and voice recorder. (#344)
* feat(playground): refactoring file uploader and voice recorder.

* Refactor voice movie

* Complete prompt movie separation

* Complete prompt movie refactoring
