+++ packages/playground-ui/src/structures/IAutoBePlaygroundBucket.ts
@@ -1,0 +1,13 @@
+import {
+  AutoBeUserMessageAudioContent,
+  AutoBeUserMessageFileContent,
+  AutoBeUserMessageImageContent,
+} from "@autobe/interface";
+
+export interface IAutoBePlaygroundBucket {
+  file: File;
+  content:
+    | AutoBeUserMessageAudioContent
+    | AutoBeUserMessageFileContent
+    | AutoBeUserMessageImageContent;
+}
