+++ packages/playground-ui/src/structures/IAutoBePlaygroundUploadConfig.ts
@@ -1,0 +1,5 @@
+export interface IAutoBePlaygroundUploadConfig {
+  supportAudio?: boolean;
+  file?: (file: File) => Promise<{ id: string }>;
+  image?: (file: File) => Promise<{ url: string }>;
+}
